D efficiency corr

+++ CodesforAnalysis/AllAnalysis.pl
@@ -2,20 +2,18 @@
 
 
 use warnings;
-
-$month = $ARGV[0]//"ALL";
+$mode =  $ARGV[0]//"0";
+$month = $ARGV[1]//"ALL";
 
 system("rm ../Histos/TOT/*");
 system("cp  ../Histos/*/*tot*  ../Histos/TOT");
 
-if($ARGV[0]=="ALL"){
-	system("rm ../Histos/TOT/TOT_tot_P1.root");
-	system ("hadd -f ../Histos/TOT/TOT_tot_P1.root ../Histos/TOT/*_tot_P1.root");
+if($ARGV[1]=="ALL"){
 	chomp (@Rootuple = `ls  ../Histos/TOT/|grep tot  |  sed s/_tot_P1.root//g`);
 
 }
-if($ARGV[0]!="ALL"){
-	@Rootuple = $ARGV[0];
+if($ARGV[1]!="ALL"){
+	@Rootuple = $ARGV[1];
 }
 
 $num_Rootuple = scalar(@Rootuple);
@@ -30,13 +28,13 @@ for($i=0;$i<$num_Rootuple-1;$i++) {
 	print OUT "\"$Rootuple[$i]\",\n";
 }
 print OUT "};\n";
-
-for($i=0;$i<$num_Rootuple;$i++) {
-	system("rm ./Final_plots/$Rootuple[$i].root");
-	print "Analyzing ".$Rootuple[$i]."...\n";
-	system("./Analysis $Rootuple[$i] 2 tot >out");
+if($ARGV[0]=="0"){
+	for($i=0;$i<$num_Rootuple;$i++) {
+		system("rm ./Final_plots/$Rootuple[$i].root");
+		print "Analyzing ".$Rootuple[$i]."...\n";
+		system("./Analysis $Rootuple[$i] 2 tot >out");
+	}
 }
-
